Switched map to MVVM
