+++ lib/main.dart
@@ -15,7 +15,7 @@ import 'package:via_mallorca/providers/tracking_provider.dart';
 import 'package:via_mallorca/screens/nearby/nearby_view.dart';
 import 'package:via_mallorca/screens/routes/routes_view.dart';
 import 'package:via_mallorca/screens/stations/stations_view.dart';
-import 'package:via_mallorca/screens/map.dart';
+import 'package:via_mallorca/screens/map/map_view.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 void main() async {
